+++ src/icons/icon.js
@@ -1,5 +1,5 @@
 export { default as NonZoonolPlaceIcon } from '@assets/icons/non_zoonol_place_icon.svg';
 export { default as CloseIcon } from '@assets/icons/close_button_icon.svg';
 export { default as zoonolPlaceIcon } from '@assets/icons/zoonol_place_icon.png';
-export { default as ButtonSheetCloseICon } from '@assets/icons/close_bottom_sheet_icon.svg';
+export { default as ButtonSheetCloseIcon } from '@assets/icons/close_bottom_sheet_icon.svg';
 export { default as CurrentLocationIcon } from '@assets/icons/current_location_icon.svg';
